Update i18n.js
The dialogue for skip intro on Netflix Japan is 「イントロをスキップ」.
The word 「紹介」, which is used in the original code, refers to "introduction" as in introducing a person or thing to someone (or presenting something/showcasing something). It doesn't have the nuance of "intro" as in English. Because of this, the word "intro" is actually borrowed from English and has been imported into Japanese as 「イントロ」(intoro).